Jump to logger call invocation
