+++ src/main/java/pl/otros/intellij/JumpToCode/model/JumpLocation.java
@@ -1,0 +1,4 @@
+package pl.otros.intellij.JumpToCode.model;
+
+public interface JumpLocation {
+}
