+++ src/main/resources/form.html
@@ -47,8 +47,8 @@
       <td><input type='text' name='l' id='lineNumber' value='120'/></td>
     </tr>
     <tr>
-      <td><label for='lineNumber'>code</label></td>
-      <td><input type='text' name='d' id='code' value=''/></td>
+      <td><label for='lineNumber'>m</label></td>
+      <td><input type='text' name='m' id='message' value=''/></td>
     </tr>
     <tr>
       <td>&nbsp;</td>
