+++ text.txt
@@ -1,0 +1,17 @@
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Unknown search mode <Random String>"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Can't load version of running OLV"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Problem with loading automatic markers, filter or log importers: "
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="OtrosLogViewer is already running, params send using requestAction"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Starting application"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Initializing look and feel: "
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Cannot initialize main frame"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Loading common configuration from "
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Can't load configuration, creating new "
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Common configuration file do not exist"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Creating user specific OtrosLogViewer configuration directory "
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Can't load user configuration from %s: %s"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Can't load user configuration from %s: %s"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Loading plugins"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Can't initialize plugins "
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Can't update search highlight"
+sleep 0.3;http http://localhost:15987/ o==jump p==pl.otros.logview.gui c==LogViewMainFrame m=="Error: "
